fix: shuffle users in search page and all-users page
#140
